--- bootcamp/src/main/java/com/globant/controler/ClientController.java
@@ -1,26 +1,0 @@
-package com.globant.controler;
-
-import java.util.List;
-
-public class ClientController implements Controller{
-
-    public Object create( Object object ) {
-        return null;
-    }
-
-    public Object read( int id ) {
-        return null;
-    }
-
-    public Object update( Object object ) {
-        return null;
-    }
-
-    public void delete( int id ) {
-
-    }
-
-    public List readAll() {
-        return null;
-    }
-}
