--- bootcamp/src/main/java/com/globant/dao/ItemDAO.java
@@ -1,28 +1,0 @@
-package com.globant.dao;
-
-import com.globant.service.Service;
-import com.globant.model.Item;
-
-import java.util.List;
-
-public class ItemDAO implements Service <Item> {
-    public Item create( Item object ) {
-        return null;
-    }
-
-    public Item read( int id ) {
-        return null;
-    }
-
-    public Item update( Item object ) {
-        return null;
-    }
-
-    public void delete( int id ) {
-
-    }
-
-    public List<Item> readAll() {
-        return null;
-    }
-}
