--- bootcamp/src/main/java/com/globant/dao/OrderDAO.java
@@ -1,28 +1,0 @@
-package com.globant.dao;
-
-import com.globant.service.Service;
-import com.globant.model.Client;
-
-import java.util.List;
-
-public class OrderDAO implements Service <Client> {
-    public Client create( Client object ) {
-        return null;
-    }
-
-    public Client read( int id ) {
-        return null;
-    }
-
-    public Client update( Client object ) {
-        return null;
-    }
-
-    public void delete( int id ) {
-
-    }
-
-    public List<Client> readAll() {
-        return null;
-    }
-}
