--- bootcamp/src/main/java/com/globant/dao/PaymentDAO.java
@@ -1,28 +1,0 @@
-package com.globant.dao;
-
-import com.globant.service.Service;
-import com.globant.model.Payment;
-
-import java.util.List;
-
-public class PaymentDAO implements Service <Payment> {
-    public Payment create( Payment object ) {
-        return null;
-    }
-
-    public Payment read( int id ) {
-        return null;
-    }
-
-    public Payment update( Payment object ) {
-        return null;
-    }
-
-    public void delete( int id ) {
-
-    }
-
-    public List<Payment> readAll() {
-        return null;
-    }
-}
