--- bootcamp/src/main/java/com/globant/service/PaymentService.java
@@ -1,30 +1,0 @@
-package com.globant.service;
-
-import com.globant.dao.PaymentDAO;
-import com.globant.dto.PaymentDTO;
-
-import java.util.List;
-
-public class PaymentService implements Service<PaymentDTO>{
-
-
-    public PaymentDTO create( PaymentDTO object ) {
-        return null;
-    }
-
-    public PaymentDTO read( int id ) {
-        return null;
-    }
-
-    public PaymentDTO update( PaymentDTO object ) {
-        return null;
-    }
-
-    public void delete( int id ) {
-
-    }
-
-    public List <PaymentDTO> readAll() {
-        return null;
-    }
-}
