+++ src/main/java/com/globant/config/AppConfiguration.java
@@ -1,0 +1,10 @@
+package com.globant.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("com.globant")
+public class AppConfiguration {
+
+}
