+++ src/main/java/com/globant/controller/ItemController.java
@@ -1,0 +1,7 @@
+package com.globant.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ItemController {
+}
