+++ src/main/java/com/globant/service/ItemService.java
@@ -1,0 +1,28 @@
+package com.globant.service;
+
+import com.globant.dao.ItemDAO;
+
+import java.util.List;
+
+public class ItemService implements CRUD<ItemDAO> {
+
+    public ItemDAO create( ItemDAO object ) {
+        return null;
+    }
+
+    public ItemDAO read( int id ) {
+        return null;
+    }
+
+    public ItemDAO update( ItemDAO object ) {
+        return null;
+    }
+
+    public void delete( int id ) {
+
+    }
+
+    public List <ItemDAO> readAll() {
+        return null;
+    }
+}
