+++ src/main/java/com/globant/service/OrderService.java
@@ -1,0 +1,27 @@
+package com.globant.service;
+
+import com.globant.dto.ClientDTO;
+
+import java.util.List;
+
+public class OrderService implements CRUD<ClientDTO> {
+    public ClientDTO create( ClientDTO object ) {
+        return null;
+    }
+
+    public ClientDTO read( int id ) {
+        return null;
+    }
+
+    public ClientDTO update( ClientDTO object ) {
+        return null;
+    }
+
+    public void delete( int id ) {
+
+    }
+
+    public List<ClientDTO> readAll() {
+        return null;
+    }
+}
