+++ src/main/java/com/globant/service/PaymentService.java
@@ -1,0 +1,28 @@
+package com.globant.service;
+
+import com.globant.dto.PaymentDTO;
+
+import java.util.List;
+
+public class PaymentService implements CRUD<PaymentDTO> {
+
+    public PaymentDTO create(PaymentDTO object ) {
+        return null;
+    }
+
+    public PaymentDTO read( int id ) {
+        return null;
+    }
+
+    public PaymentDTO update( PaymentDTO object ) {
+        return null;
+    }
+
+    public void delete( int id ) {
+
+    }
+
+    public List <PaymentDTO> readAll() {
+        return null;
+    }
+}
